// src/types/settings.ts
export interface PublicSettings {
  shippingDisclaimer: string;
  bankTransferBank: string;
  bankTransferHolder: string;
  bankTransferCbu: string;
  bankTransferAlias: string;
  contactAddress: string;
  contactPhone: string;
  contactEmail: string;
  socialFacebook: string;
  socialInstagram: string;
  socialYoutube: string;
}